pain

+++ render/scale.c
@@ -11,7 +11,7 @@ void Scale_render(void) {
 	Color temp[h][w];
 	for (axis y=0; y<h; y++)
 		for (axis x=0; x<w; x++)
-			temp[y][x] = grp[Menu_zoomY+y+8][Menu_zoomX+x+8];
+			temp[y][x] = grp[(int)Menu_zoomY+y+8][(int)Menu_zoomX+x+8];
 	for (axis y=0;y<H;y++) {
 		for (axis x=0; x<W; x++) {
 			grp[y+8][x+8] = temp[y/zoom][x/zoom];
